Adds hash tests unique_coroutine_handle, is noexcept move constructible
